stress-sctp: pass int64_t types for SCTP_MAXSEG and SCTP_MAX_BURST sock opts
Modern use of these sock options should use sctp_assoc_value which is an 8 byte
struct. If sctp_assoc_value is not available, use a 64 bit integer type.

+++ stress-sctp.c
@@ -174,7 +174,7 @@ static void stress_sctp_sockopts(const int fd)
 		static bool once = false;
 
 		if (once == false) {
-			STRESS_SCTP_SOCKOPT(SCTP_MAXSEG, int)
+			STRESS_SCTP_SOCKOPT(SCTP_MAXSEG, int64_t)
 			once = true;
 		}
 	}
@@ -209,7 +209,7 @@ static void stress_sctp_sockopts(const int fd)
 		static bool once = false;
 
 		if (once == false) {
-			STRESS_SCTP_SOCKOPT(SCTP_MAX_BURST, unsigned int)
+			STRESS_SCTP_SOCKOPT(SCTP_MAX_BURST, uint64_t)
 			once = true;
 		}
 	}
